Handle row updates that modify primary key
Thanks to James Cheng for pointing out this edge case. Relational
databases allow a row's primary key to be updated, just like any other
column. When mapping this to a key-value model, which doesn't have the
concept of modifying a key, an update to a primary key needs to be
transformed into a delete of the old primary key, and an insert of the
new key.

Fortunately, this is easy to support within the existing protocol.