Use per-tuple memory context in the export SRF.
The bottledwater_export() set returning function should allocate all
memory it needs to process a tuple in the dedicated context to avoid
filling up executor state or the SPI proc context (whichever happens
to be the current).

The result to be returned should be allocated on the said memory
context and it should be reset every time *before* producing output
for the next tuple.