Build client and postgres images from tmp directory...
This allows us to set a .dockerignore for the project root to exclude
the tmp directory (and thus the built tarballs) from the main
Dockerfile.build, but bypass that .dockerignore (i.e. include the
tarballs in the build context) for the postgres and client image builds.

Without this, subsequent runs of Dockerfile.build always receive the
tarballs built by the previous run in their Docker build context, so the
build context is always different and Docker never caches the build.