"flush" wal_pos reported in keepalive unless we have pending txns...
The server may generate WAL that doesn't correspond to a transaction
commit, e.g.:

 * automatic checkpoints driven by checkpoint_timeout
   (http://www.postgresql.org/docs/current/static/runtime-config-wal.html#GUC-CHECKPOINT-TIMEOUT)
 * transactions that roll back

Currently we only update our "flush lsn" when we see a COMMIT, so if
enough of this non-committed-transaction activity occurs before we see a
COMMIT, the server may run out of WAL space and crash.

The server's regular keepalive messages are already informing us of the
updated wal_pos resulting from this activity.  So this commit makes us
acknowledge it as "flushed" (even though we didn't actually do anything
about it), provided that we aren't currently waiting for Kafka to
acknowledge a transaction in flight (in which case acknowledging the
subsequent activity would could result in us losing the transaction in
flight, in case of a restart).