Use container hostnames instead of link environment variables...
Docker populates environment variables for each container link, but
those environment variables are deprecated (as are links themselves):
see https://docs.docker.com/compose/link-env-deprecated/.  This stops
relying on those environment variables in our Docker setup, and instead
addresses containers by hostname.

For links, Docker populates /etc/hosts with suitable hostnames.  The
hostname method should also work when we move from links to Docker
networks, where Docker provides DNS resolution for container names.

This was mainly prompted (see #79) by a change in the confluent/tools
Docker image, which used to provide a wrapper script that read the
environment variables and started up the tools with the appropriate
arguments.  Now that script is no longer available as an entrypoint
(only as a script that generates aliases), so we need to pass the
arguments ourselves.