Update bottledwater.c
In send_kafka_msg, it seems that rdkafka doesn't free the payload (it is 'val' in our situation ) when it is done with the payload. That why I propose that we change from RD_KAFKA_MSG_F_FREE to RD_KAFKA_MSG_F_COPY. Then we manually free val and key.