Don't ignore errors creating Avro schema...
If avro-c is unable to create the schema (e.g. due to an invalid
identifier), it sets avro_strerror and returns NULL (or an error code if
it was a schema update that failed). We weren't checking for this, so we
would just store a NULL or invalid schema, pass it later to a function
that didn't expect that, and segfault.  This way we fail faster, and
with a more informative error message.