Sanitise strings before using them as identifiers for Avro schemas...
This is necessary because Postgres is much more permissive than Avro
about what characters an identifier can contain: e.g. "user.friend" is a
valid Postgres table name, but not a valid Avro schema name.

This encoding has some gotchas, notably that it is unspecified for
non-ASCII characters and that it is not completely unambiguous.